requirements.txt
aiohttp
beautifulsoup4
pyYAML